Added the __init__.py and __main__.py file.
The __init__.py file was added which contains the
source code.

The 4 mainfunctions are:
    - get_letter_on_pos()
    - get_letters_from_pos_to_pos()
    - get_fields()
    - get_fields_new_separator()

All functions are well documented in the source code. The code also
contains a lot of comments to make adjustments easier later on.

The __main__.py file creates a commandline version which only shows
help at the moment.